# Optimize PGBK table to only update cache when there is a large enough size change. #21250