--- a/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/PrecombineGroupingTable.java
+++ b/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/PrecombineGroupingTable.java
@@ -128,8 +128,7 @@ public class PrecombineGroupingTable<K, InputT, AccumT>
   private final AtomicLong maxWeight;
   private long weight;
   private final boolean isGloballyWindowed;
-  private long checkFlushCounter;
-  private long checkFlushLimit = -5;
+  private long lastWeightForFlush;
 
   private static final class Key implements Weighted {
     private static final Key INSTANCE = new Key();
@@ -407,12 +406,9 @@ public class PrecombineGroupingTable<K, InputT, AccumT>
           return tableEntry;
         });
 
-    if (checkFlushCounter++ < checkFlushLimit) {
-      return;
-    } else {
-      checkFlushLimit = Math.min(checkFlushLimit + 1, 25);
-      checkFlushCounter = 0;
+    if (Caches.shouldUpdateOnSizeChange(lastWeightForFlush, weight)) {
       flushIfNeeded(receiver);
+      lastWeightForFlush = weight;
     }
   }